tests/wpt/tests/html/canvas/offscreen/layers/2d.layer.ctm.getTransform.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.layer.ctm.getTransform</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.layer.ctm.getTransform</h1>
<p class="desc">Tests getTransform inside layers.</p>


<script>
var t = async_test("Tests getTransform inside layers.");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.translate(10, 20);
  ctx.beginLayer();
  ctx.scale(2, 3);
  const m = ctx.getTransform();
  assert_array_equals([m.a, m.b, m.c, m.d, m.e, m.f], [2, 0, 0, 3, 10, 20]);
  ctx.endLayer();
  t.done();

});
</script>
